requirements.txt
numpy
scipy
torch       
paho-mqtt
pyttsx3
SpeechRecognition
openai
tk 
pyaudio
# use pip3 install --user -r requirements.txt